requirements.txt
asn1crypto==1.4.0
attrs==19.3.0
Automat==0.8.0
base58==2.1.0
bip-utils==2.0.2
bitcoin==1.1.42
blinker==1.4
btclib==2021.1
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.6
Click==7.0
cloud-init==21.1
codetiming==1.3.0
coincurve==15.0.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
crcmod==1.7
cryptography==2.8
dataclasses-json==0.5.6
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
ecdsa==0.17.0
ed25519-blake2b==1.4
entrypoints==0.3
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
marshmallow==3.13.0
marshmallow-enum==1.5.1
more-itertools==4.2.0
mypy-extensions==0.4.3
netifaces==0.10.4
oauthlib==3.1.0
pexpect==4.6.0
pprintpp==0.4.0
py-sr25519-bindings==0.1.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycryptodome==3.10.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
pysha3==1.0.2
python-apt==2.0.0+ubuntu0.20.4.4
python-bitcointx==1.1.2
python-debian===0.1.36ubuntu1
PyYAML==5.3.1
requests==2.26.0
requests-unixsocket==0.2.0
scalecodec==1.0.1
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sos==4.1
ssh-import-id==5.10
systemd-python==234
Twisted==18.9.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
zipp==1.0.0
zope.interface==4.7.1
